+++ src/main/java/lotto/domain/lotto/CreateLottoNumbersStrategy.java
@@ -1,4 +1,6 @@
-package lotto.domain;
+package lotto.domain.lotto;
+
+import lotto.domain.lotto.LottoNumber;
 
 import java.util.List;
