+++ src/main/java/lotto/domain/reward/Reward.java
@@ -1,4 +1,6 @@
-package lotto.domain;
+package lotto.domain.reward;
+
+import lotto.domain.Money;
 
 import java.util.Arrays;
