+++ src/test/java/lotto/domain/LottoBuyerTest.java
@@ -1,5 +1,7 @@
 package lotto.domain;
 
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.Lottos;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
@@ -8,8 +10,8 @@ import lotto.dto.LottoStatisticsResult;
 
 import static org.assertj.core.api.Assertions.catchThrowable;
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
-import static lotto.domain.Lotto.LOTTO_NUMBER_MUST_NOT_BE_BLANK;
-import static lotto.domain.Lotto.LOTTO_NUMBER_SIZE_NOT_VALID;
+import static lotto.domain.lotto.Lotto.LOTTO_NUMBER_MUST_NOT_BE_BLANK;
+import static lotto.domain.lotto.Lotto.LOTTO_NUMBER_SIZE_NOT_VALID;
 import static lotto.domain.LottoBuyer.MONEY_MUST_GRATE_THEN_ZERO_TO_BUY_LOTTO;
 import static lotto.domain.Money.MONEY_MUST_NOT_BE_NEGATIVE;
