+++ src/test/java/lotto/domain/MoneyTest.java
@@ -1,5 +1,6 @@
 package lotto.domain;
 
+import lotto.domain.lotto.Lotto;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
