+++ src/test/java/lotto/domain/lotto/CreateRandomNumbersStrategyTest.java
@@ -1,5 +1,9 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
+import lotto.domain.lotto.CreateLottoNumbersStrategy;
+import lotto.domain.lotto.CreateRandomNumbersStrategy;
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.LottoNumber;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
