+++ src/test/java/lotto/domain/lotto/LottoMockFactory.java
@@ -1,5 +1,7 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.LottoNumber;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
